backend/requirements.txt
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
cachetools==5.3.1
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
cryptography==41.0.3
cssselect==1.2.0
d20==1.1.2
Django==4.1
django-cors-headers==4.2.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
lark-parser==0.9.0
lxml==4.9.3
mssql-django==1.3
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
pdfminer==20191125
pdfminer.six==20221105
pdfquery==0.4.3
pycparser==2.21
pycryptodome==3.18.0
PyMySQL==1.1.0
pyodbc==4.0.39
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3
roman==4.1
six==1.16.0
sqlparse @ file:///C:/b/abs_83jn6owfpc/croot/sqlparse_1668793499314/work
tzdata==2023.3
